% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clinical_datasets.R
\docType{data}
\name{clinical}
\alias{ACC.clinical}
\alias{ACC.clinical,BLCA.clinical,BRCA.clinical,CESC.clinical,CHOL.clinical,COAD.clinical,COADREAD.clinical,DLBC.clinical,ESCA.clinical,FPPP.clinical,GBM.clinical,GBMLGG.clinical,HNSC.clinical,KICH.clinical,KIPAN.clinical,KIRC.clinical,KIRP.clinical,LAML.clinical,LGG.clinical,LIHC.clinical,LUAD.clinical,LUSC.clinical,MESO.clinical,OV.clinical,PAAD.clinical,PCPG.clinical,READ.clinical,SKCM.clinical,STAD.clinical,STES.clinical,TGCT.clinical,THCA.clinical,THYM.clinical,UCEC.clinical,UCS.clinical,UVM.clinical}
\alias{BLCA.clinical}
\alias{BRCA.clinical}
\alias{CESC.clinical}
\alias{CHOL.clinical}
\alias{COAD.clinical}
\alias{COADREAD.clinical}
\alias{DLBC.clinical}
\alias{ESCA.clinical}
\alias{FPPP.clinical}
\alias{GBM.clinical}
\alias{GBMLGG.clinical}
\alias{HNSC.clinical}
\alias{KICH.clinical}
\alias{KIPAN.clinical}
\alias{KIRC.clinical}
\alias{KIRP.clinical}
\alias{LAML.clinical}
\alias{LGG.clinical}
\alias{LIHC.clinical}
\alias{LUAD.clinical}
\alias{LUSC.clinical}
\alias{MESO.clinical}
\alias{OV.clinical}
\alias{PAAD.clinical}
\alias{PCPG.clinical}
\alias{READ.clinical}
\alias{SKCM.clinical}
\alias{STAD.clinical}
\alias{STES.clinical}
\alias{TGCT.clinical}
\alias{THCA.clinical}
\alias{THYM.clinical}
\alias{UCEC.clinical}
\alias{UCS.clinical}
\alias{UVM.clinical}
\alias{clinical}
\title{Clinical datasets from TCGA project}
\source{
\url{http://gdac.broadinstitute.org/}
}
\usage{
ACC.clinical

BLCA.clinical

BRCA.clinical

CESC.clinical

CHOL.clinical

COAD.clinical

COADREAD.clinical

DLBC.clinical

ESCA.clinical

FPPP.clinical

GBM.clinical

GBMLGG.clinical

HNSC.clinical

KICH.clinical

KIPAN.clinical

KIRC.clinical

KIRP.clinical

LAML.clinical

LGG.clinical

LIHC.clinical

LUAD.clinical

LUSC.clinical

MESO.clinical

OV.clinical

PAAD.clinical

PCPG.clinical

READ.clinical

SKCM.clinical

STAD.clinical

STES.clinical

TGCT.clinical

THCA.clinical

THYM.clinical

UCEC.clinical

UCS.clinical

UVM.clinical
}
\value{
Data frames with clinical data.
}
\description{
Package provides clinical datasets from The Cancer Genome Atlas Project for all cohorts types from \url{http://gdac.broadinstitute.org/}.
Data were downloaded using \link{RTCGA-package} and contain snapshots for the date: \code{2015-08-21}. The process is described here: \url{http://mi2-warsaw.github.io/RTCGA.data/}.
Use cases, examples and information about datasets in \pkg{RTCGA.data} family can be found here: \code{browseVignettes("RTCGA")}. Clinical data format is explained here \url{https://wiki.nci.nih.gov/display/TCGA/Clinical+Data+Overview}.
}
\details{
\code{browseVignettes("RTCGA")}
}
\keyword{datasets}

